+++ dropwizard-hibernate/src/main/java/io/dropwizard/hibernate/UnitOfWorkApplicationListener.java
@@ -1,7 +1,6 @@
 package io.dropwizard.hibernate;
 
 import org.glassfish.jersey.server.internal.process.MappableException;
-import org.glassfish.jersey.server.model.Resource;
 import org.glassfish.jersey.server.model.ResourceMethod;
 import org.glassfish.jersey.server.monitoring.ApplicationEvent;
 import org.glassfish.jersey.server.monitoring.ApplicationEventListener;
@@ -10,9 +9,11 @@ import org.glassfish.jersey.server.monitoring.RequestEventListener;
 import org.hibernate.SessionFactory;
 
 import javax.ws.rs.ext.Provider;
-import java.lang.reflect.Method;
 import java.util.HashMap;
 import java.util.Map;
+import java.util.Optional;
+import java.util.concurrent.ConcurrentHashMap;
+import java.util.concurrent.ConcurrentMap;
 
 
 /**
@@ -27,7 +28,7 @@ import java.util.Map;
 @Provider
 public class UnitOfWorkApplicationListener implements ApplicationEventListener {
 
-    private Map<Method, UnitOfWork> methodMap = new HashMap<>();
+    private ConcurrentMap<ResourceMethod, Optional<UnitOfWork>> methodMap = new ConcurrentHashMap<>();
     private Map<String, SessionFactory> sessionFactories = new HashMap<>();
 
     public UnitOfWorkApplicationListener() {
@@ -58,11 +59,11 @@ public class UnitOfWorkApplicationListener implements ApplicationEventListener {
     }
 
     private static class UnitOfWorkEventListener implements RequestEventListener {
-        private final Map<Method, UnitOfWork> methodMap;
+        private ConcurrentMap<ResourceMethod, Optional<UnitOfWork>> methodMap;
         private final UnitOfWorkAspect unitOfWorkAspect;
 
-        UnitOfWorkEventListener(Map<Method, UnitOfWork> methodMap,
-                                       Map<String, SessionFactory> sessionFactories) {
+        UnitOfWorkEventListener(ConcurrentMap<ResourceMethod, Optional<UnitOfWork>> methodMap,
+                                Map<String, SessionFactory> sessionFactories) {
             this.methodMap = methodMap;
             unitOfWorkAspect = new UnitOfWorkAspect(sessionFactories);
         }
@@ -71,9 +72,9 @@ public class UnitOfWorkApplicationListener implements ApplicationEventListener {
         public void onEvent(RequestEvent event) {
             final RequestEvent.Type eventType = event.getType();
             if (eventType == RequestEvent.Type.RESOURCE_METHOD_START) {
-                UnitOfWork unitOfWork = methodMap.get(event.getUriInfo()
-                        .getMatchedResourceMethod().getInvocable().getDefinitionMethod());
-                unitOfWorkAspect.beforeStart(unitOfWork);
+                Optional<UnitOfWork> unitOfWork = methodMap.computeIfAbsent(event.getUriInfo()
+                        .getMatchedResourceMethod(), UnitOfWorkEventListener::registerUnitOfWorkAnnotations);
+                unitOfWorkAspect.beforeStart(unitOfWork.orElse(null));
             } else if (eventType == RequestEvent.Type.RESP_FILTERS_START) {
                 try {
                     unitOfWorkAspect.afterEnd();
@@ -86,23 +87,18 @@ public class UnitOfWorkApplicationListener implements ApplicationEventListener {
                 unitOfWorkAspect.onFinish();
             }
         }
+
+        private static Optional<UnitOfWork> registerUnitOfWorkAnnotations(ResourceMethod method) {
+            UnitOfWork annotation = method.getInvocable().getDefinitionMethod().getAnnotation(UnitOfWork.class);
+            if (annotation == null) {
+                annotation = method.getInvocable().getHandlingMethod().getAnnotation(UnitOfWork.class);
+            }
+            return Optional.ofNullable(annotation);
+        }
     }
 
     @Override
     public void onEvent(ApplicationEvent event) {
-        if (event.getType() == ApplicationEvent.Type.INITIALIZATION_APP_FINISHED) {
-            for (Resource resource : event.getResourceModel().getResources()) {
-                for (ResourceMethod method : resource.getAllMethods()) {
-                    registerUnitOfWorkAnnotations(method);
-                }
-
-                for (Resource childResource : resource.getChildResources()) {
-                    for (ResourceMethod method : childResource.getAllMethods()) {
-                        registerUnitOfWorkAnnotations(method);
-                    }
-                }
-            }
-        }
     }
 
     @Override
@@ -110,16 +106,5 @@ public class UnitOfWorkApplicationListener implements ApplicationEventListener {
         return new UnitOfWorkEventListener(methodMap, sessionFactories);
     }
 
-    private void registerUnitOfWorkAnnotations(ResourceMethod method) {
-        UnitOfWork annotation = method.getInvocable().getDefinitionMethod().getAnnotation(UnitOfWork.class);
 
-        if (annotation == null) {
-            annotation = method.getInvocable().getHandlingMethod().getAnnotation(UnitOfWork.class);
-        }
-
-        if (annotation != null) {
-            this.methodMap.put(method.getInvocable().getDefinitionMethod(), annotation);
-        }
-
-    }
 }